refactor(4a): reduce cockpitd.ts to host (139 LOC); add daemon/start.ts (startDaemon)
daemon/start.ts contains all assembly logic: createProbes, createDelivery,
createDaemon + all callbacks, buildHealth/gatherSnapshotInputs, boot-recovery
IIFE, createServer, all timers, and the DaemonHandle return. No concrete driver
imports (headless-launcher/codex-driver/cmux-daemon-cmux stay in the host).

cockpitd.ts becomes the thin host: builds ctx, constructs CodexInteractiveDriver /
OpencodeSseBridge / CmuxEventsBridge / DaemonCmux, assembles launchHeadless, and
calls startDaemon(ctx, opts, PKG_VERSION). Existing startCockpitd name is preserved
for test compatibility.
